+++ app/src/main/java/com/NumCo/numberconverter/Numerals/Decimal.java
@@ -4,21 +4,21 @@ public class Decimal extends Numeral {
 
     private String decimal;
 
-    public Decimal(String decimal){
+    public Decimal(String decimal) {
         super(decimal);
         this.decimal = decimal;
     }
 
-    public String toHex(){
-        return Integer.toHexString(Integer.parseInt(decimal));
+    public String toHex() {
+        return Long.toHexString(Integer.parseInt(decimal));
     }
 
-    public String toBin(){
-        return Integer.toBinaryString(Integer.parseInt(decimal));
+    public String toBin() {
+        return Long.toBinaryString(Integer.parseInt(decimal));
     }
 
-    public String toOct(){
-        return Integer.toOctalString(Integer.parseInt(decimal));
+    public String toOct() {
+        return Long.toOctalString(Integer.parseInt(decimal));
     }
 
 }
