+++ app/src/main/java/com/NumCo/numberconverter/Numerals/Octal.java
@@ -4,21 +4,21 @@ public class Octal extends Numeral {
 
     private String octal;
 
-    public Octal(String octal){
+    public Octal(String octal) {
         super(octal);
         this.octal = octal;
     }
 
-    public String toHex(){
-        return Integer.toHexString(Integer.parseInt(octal, 8));
+    public String toHex() {
+        return Long.toHexString(Long.parseLong(octal, 8));
     }
 
-    public String toBin(){
-        return Integer.toBinaryString(Integer.parseInt(octal, 8));
+    public String toBin() {
+        return Long.toBinaryString(Long.parseLong(octal, 8));
     }
 
-    public String toDec(){
-        return Integer.toString(Integer.parseInt(octal, 8));
+    public String toDec() {
+        return Long.toString(Long.parseLong(octal, 8));
     }
 
 }
